requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
python-decouple==3.8
psycopg2-binary==2.9.9
python-dateutil==2.8.2
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0